backend/requirements.txt
Django
python-decouple
django-debug-toolbar
requests
beautifulsoup4
pdfplumber
mysqlclient
pymysql
cryptography
django-ninja
django-cors-headers
django-ninja-extra
google-auth
pytz
tzdata